refactor(sortingHelpers): move helper functions to subdirectory and update imports
